scrapy_project/requirements.txt
torch
torchvision
torchaudio
transformers
ultralytics
pillow
requests